backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
python-dotenv==1.0.0
PyMuPDF==1.23.8
faiss-cpu==1.7.4
google-generativeai==0.3.2
numpy==1.24.3
email-validator
passlib[bcrypt]
python-jose[cryptography]
beautifulsoup4==4.12.2
pgvector